+++ CHANGELOG.rst
@@ -27,8 +27,9 @@ Changed
 - narhwalify SetValueTransformer `#398 <https://github.com/azukds/tubular/issues/398>_`
 - narwhalified DateDifferenceTransformer. `#376 <https://github.com/azukds/tubular/issues/376>_`
 - narwhalified DateDiffLeapYearTransformer.
-- placeholder
+- narwhalified MappingTransformer `#374 <https://github.com/azukds/tubular/issues/374>_`
 - placeholder 
+- placeholder
 - placeholder
 
 1.4.2 (18/03/2025)